libcontainer/user: move userFromOS,groupFromOS to Windows files
These are only used on Windows, causing the deadcode linter to complain:

    Error: `userFromOS` is unused (deadcode)
    Error: `groupFromOS` is unused (deadcode)
